+++ kudzu-core/src/commonMain/kotlin/com/copperleaf/kudzu/parser/text/LiteralTokenParser.kt
@@ -63,17 +63,19 @@ class LiteralTokenParser(
     private fun doParse(input: ParserContext): Pair<ParserContext?, String> {
         var remaining = input
         var nextChar: Char
-        var token = ""
+        val tokenBuilder = StringBuilder()
         var index = 0
 
         while (!remaining.isEmpty()) {
             nextChar = remaining.next()
             if (!isValidChar(index, nextChar)) break
 
-            token += nextChar
+            tokenBuilder.append(nextChar)
             remaining = remaining.remaining()
             index++
         }
+
+        val token = tokenBuilder.toString()
 
         return (if (token == expected) remaining else null) to token
     }